Extract CSipHasher to it's own file in crypto/ directory.
Summary:
This is a move-only commit with the exception of changes to includes.

backported from core PR14074 ( 4fb789e9b )

Test Plan: `make check`

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Subscribers: Fabien, jasonbcox

Maniphest Tasks: T589

Differential Revision: https://reviews.bitcoinabc.org/D2800
