+++ src/bench/crypto_hash.cpp
@@ -10,6 +10,7 @@
 #include "crypto/sha1.h"
 #include "crypto/sha256.h"
 #include "crypto/sha512.h"
+#include "crypto/siphash.h"
 #include "hash.h"
 #include "random.h"
 #include "uint256.h"
