+++ src/test/hash_tests.cpp
@@ -1,10 +1,11 @@
-// Copyright (c) 2013-2016 The Bitcoin Core developers
+// Copyright (c) 2013-2018 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "hash.h"
 #include "test/test_bitcoin.h"
 #include "utilstrencodings.h"
+#include "crypto/siphash.h"
 
 #include <vector>
