+++ src/txmempool.h
@@ -13,6 +13,7 @@
 #include "primitives/transaction.h"
 #include "random.h"
 #include "sync.h"
+#include "crypto/siphash.h"
 
 #include <boost/multi_index/hashed_index.hpp>
 #include <boost/multi_index/ordered_index.hpp>
